requirements.txt
annotated-types==0.7.0
audioread==3.0.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
coloredlogs==15.0.1
decorator==5.2.1
flatbuffers==25.2.10
greenlet==3.2.3
humanfriendly==10.0
idna==3.10
joblib==1.5.1
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.44.0
mpmath==1.3.0
msgpack==1.1.1
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
packaging==25.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
pyacoustid==1.3.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.0
soundfile==0.13.1
soxr==0.5.0.post1
speechmos==0.0.1.1
sqlalchemy==2.0.41
sympy==1.14.0
threadpoolctl==3.6.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
